Introduce `LazyIvyNgElementStrategy` that registers a dummy/shell Custom Element for a given tag name and lazily initializes the underlying Angular component only when the data binding changes.
On the server, the same strategy bootstraps the component eagerly(since the ComponentType is passed in instead of the tag name) and serializes the Input properties using JSON.stringify. On the client the strategy uses these initial DOM attributes for hydrating the component properties before initial render - So that rehydration can properly line up.

Any new property changes are applied after the initial render and a change detection is trigerred.

This allows the client to lazily and progressively bootstrap elements only as needed based on data binding changes. A following commit will do a similar thing, using events sent to an Angular Element as a signal to lazily bootstrap a component.