Support client side updates with the rehydration renderer.
Identify the case when an element is rendered only in the client-side(Does not have attribute "_s" set) and immediately loads the component instead of waiting for event or property changes.

Also cleaned up the webpack named chunks to automatically include elements/*.ts but exclude test files which would be elements/*.spec.ts.