add support for fetching initial data in a page/component.
The data is fetched/resolved before the component is rendered - On the server this removes the need for change detection and on client this allows for the component to rehydrate properly in place.

A component can optionally expose a getInitialInputs static method to resove/fetch any extra inputs.

Any data fetched in this phase on the server is automatically transferred to the client through the DOM so that the duplicate fetch is avoided.

Needs some more work around error handling and propagation.